[+] Lista de Downloads (WIP).

+++ src/com/ses/zxdb/dao/MachineType.kt
@@ -3,4 +3,4 @@ package com.ses.zxdb.dao
 import com.ses.sql.Table
 
 @Table("machinetypes")
-class MachineType: EnumerationTable<Byte>()
+class MachineType: EnumerationTable<Int>()
